+++ lightning-commons/lightning-commons-client-core/src/main/java/org/tbk/lightning/client/common/core/LightningCommonClient.java
@@ -1,0 +1,31 @@
+package org.tbk.lightning.client.common.core;
+
+import org.tbk.lightning.client.common.core.proto.*;
+import reactor.core.publisher.Mono;
+
+public interface LightningCommonClient {
+    Mono<CommonInfoResponse> info(CommonInfoRequest request);
+
+    Mono<CommonConnectResponse> connect(CommonConnectRequest request);
+
+    Mono<CommonNewAddressResponse> newAddress(CommonNewAddressRequest request);
+
+    Mono<CommonCreateInvoiceResponse> createInvoice(CommonCreateInvoiceRequest request);
+
+    Mono<CommonListPeersResponse> listPeers(CommonListPeersRequest request);
+
+    Mono<CommonOpenChannelResponse> openChannel(CommonOpenChannelRequest request);
+
+    Mono<CommonListUnspentResponse> listUnspent(CommonListUnspentRequest request);
+
+    Mono<CommonListPeerChannelsResponse> listPeerChannels(CommonListPeerChannelsRequest request);
+
+    Mono<CommonPayResponse> pay(CommonPayRequest request);
+
+    Mono<CommonLookupInvoiceResponse> lookupInvoice(CommonLookupInvoiceRequest request);
+
+    Mono<CommonLookupPaymentResponse> lookupPayment(CommonLookupPaymentRequest request);
+
+    Mono<CommonQueryRouteResponse> queryRoutes(CommonQueryRouteRequest request);
+
+}